update python bindings
