+++ imu/__init__.py
@@ -3,7 +3,7 @@
 from .bindings import (
     PyHexmoveImuData as HexmoveImuData,
     PyHexmoveImuReader as HexmoveImuReader,
-)
-from .bindings_hiwonder import (
     PyHiwonderImu as HiwonderImu,
 )
+
+__all__ = ['HexmoveImuData', 'HexmoveImuReader', 'HiwonderImu']
