--- imu/bindings_hiwonder/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "bindings_hiwonder"
-version = "0.1.0"
-edition = "2021"
-authors = ["Aaron Xie <aaron@kscale.dev>"]
-description = "Python bindings for Hiwonder IMU"
-license = "MIT"
-
-[lib]
-name = "bindings_hiwonder"
-crate-type = ["cdylib", "rlib"]
-
-[dependencies]
-pyo3 = { version = ">= 0.21.0", features = ["extension-module"] }
-pyo3-stub-gen = ">= 0.6.0"
-hiwonder = { path = "../hiwonder" }
