--- imu/bindings_hiwonder/src/bin/stub_gen.rs
@@ -1,7 +1,0 @@
-use pyo3_stub_gen::Result;
-
-fn main() -> Result<()> {
-    let stub = bindings_hiwonder::stub_info()?;
-    stub.generate()?;
-    Ok(())
-}
