--- imu/hiwonder/__init__.py
@@ -1,5 +1,0 @@
-"""Defines the top-level API for the Hiwonder IMU package."""
-
-from bindings_hiwonder import HiwonderImu  # Changed from relative import
-
-__all__ = ['HiwonderImu']
